requirements.txt
alabaster==0.7.16
alembic==1.12.0
amqp==5.1.1
anyio==4.4.0
APScheduler==3.10.1
arabic-reshaper==3.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asn1crypto==1.5.1
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Authlib==1.3.1
autobahn==24.4.2
Automat==22.10.0
azure-common==1.1.28
azure-core==1.31.0
azure-identity==1.17.1
azure-mgmt-core==1.4.0
azure-mgmt-rdbms==10.1.0
azure-mgmt-resource==23.1.1
azure-mgmt-subscription==3.1.1
Babel==2.13.0
backports.tarfile==1.2.0
bcrypt==4.2.0
beautifulsoup4==4.9.3
bidict==0.22.1
billiard==3.6.4.0
bleach==6.1.0
blinker==1.6.3
boto3==1.35.25
botocore==1.35.25
Brotli==1.1.0
bs4==0.0.1
build==0.9.0
cachetools==5.3.1
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.2.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
comm==0.2.2
constantly==23.10.4
crispy-bootstrap3==2024.1
crispy-bootstrap4==2023.1
cryptography==43.0.1
cssselect2==0.7.0
daphne==4.0.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Django==5.0.4
django-appconf==1.0.6
django-celery-results==2.4.0
django-ckeditor==6.5.1
django-cleanup==5.2.0
django-clearcache==1.2.1
django-crispy-forms==2.0
django-db-connection-pool==1.2.5
django-debug-toolbar==4.2.0
django-filter==2.4.0
django-js-asset==2.0.0
django-mptt==0.15.0
django-postgrespool2==2.0.5
django-qr-code==2.2.0
django-redis==5.4.0
django-rest-framework-datatables-filter==0.0.2
django-rest-multiple-models==2.1.3
django-select2==8.1.2
django-simple-history==3.0.0
django-storages==1.11.1
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
djangorestframework-datatables==0.6.0
djangorestframework-datatables-editor==0.3.3
dnspython==2.4.2
docutils==0.20.1
email-validator==2.0.0.post2
et-xmlfile==1.1.0
eventlet==0.36.1
executing==2.0.1
falcon==3.1.3
fastjsonschema==2.19.1
Flask==3.0.3
flask-babel==4.0.0
Flask-Compress==1.14
Flask-Gravatar==0.5.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-Paranoid==0.3.0
Flask-Principal==0.4.0
Flask-Security-Too==5.5.2
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
forex-python==1.8
fqdn==1.5.1
future==0.18.2
google-api-core==2.12.0
google-api-python-client==2.103.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.61.0
greenlet==3.0.0
h11==0.14.0
html5lib==1.1
httpagentparser==1.9.5
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
hyperlink==21.0.0
idna==2.10
imagesize==1.4.1
importlib-metadata==6.8.0
importlib_resources==6.4.0
incremental==22.10.0
ipykernel==6.29.4
ipython==8.25.0
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.1.2
jaraco.classes==3.3.0
jaraco.context==6.0.1
jaraco.functools==4.0.2
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.2
jmespath==0.10.0
json5==0.9.25
jsonformatter==0.3.2
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
keyring==25.4.1
kombu==5.2.4
ldap3==2.9.1
libgravatar==1.0.4
lxml==5.3.0
Mako==1.2.4
Markdown==3.3.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
more-itertools==10.1.0
msal==1.31.0
msal-extensions==1.2.0
msgpack==1.0.8
msrest==0.7.1
multipart==0.2.4
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
openpyxl==3.1.2
oscrypto==1.3.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
paramiko==3.5.0
parso==0.8.4
passlib==1.7.4
pdf2image==1.17.0
pdfservices-sdk==4.0.0
pep517==0.13.0
pexpect==4.9.0
#pgadmin4==8.12
Pillow==9.4.0
pip-install==1.3.5
platformdirs==4.2.2
polling==0.3.2
polling2==0.5.0
portalocker==2.8.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==4.24.4
psutil==6.0.0
#psycopg==3.2.1
#psycopg-binary==3.2.1
#psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.14.0
pyHanko==0.17.2
pyhanko-certvalidator==0.20.1
pyheif==0.7.1
pyheif-pillow-opener==0.2.0
PyJWT==2.4.0
PyNaCl==1.5.0
pyodbc==5.1.0
pyOpenSSL==24.2.1
pyotp==2.9.0
pyparsing==3.0.9
pypdf==3.7.1
PyPDF2==1.26.0
pypng==0.20220715.0
PySocks==1.7.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==0.18.0
python-engineio==4.8.0
python-json-logger==2.0.7
python-socketio==5.9.0
pytz==2024.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==26.0.3
qrcode==7.4.2
redis==5.0.1
referencing==0.35.1
reportlab==3.6.12
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
s3transfer==0.10.2
SecretStorage==3.3.3
segno==1.3.3
selenium==4.18.1
Send2Trash==1.8.3
sentry-sdk==1.39.1
service-identity==23.1.0
shellingham==1.5.4
simple-websocket==1.0.0
simplejson==3.19.1
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
speaklater3==1.4
Sphinx==7.3.7
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.22
sqlparams==6.0.1
sqlparse==0.4.1
sshtunnel==0.4.0
stack-data==0.6.3
svglib==1.5.1
telnetlib3==2.0.4
terminado==0.18.1
tinycss2==1.2.1
toml==0.10.2
tornado==6.4.1
traitlets==5.14.3
trio-websocket==0.11.1
Twisted==23.10.0
txaio==23.1.1
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.10.0
tzdata==2023.3
tzlocal==4.3
ua-parser==0.18.0
Unidecode==1.3.8
uri-template==1.3.0
uritemplate==4.1.1
uritools==4.0.1
urllib3==1.26.13
user-agents==2.2.0
uvicorn==0.27.1
validate-email==1.3
vine==5.0.0
wcwidth==0.2.6
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.0.0
wsproto==1.2.0
WTForms==3.1.2
xhtml2pdf==0.2.9
xlrd==2.0.1
XlsxWriter==3.0.1
xmltodict==0.13.0
zipp==3.17.0
zope.interface==6.1
